Update Newtonsoft.Json (13.0.1 has security vulnerability)
BlackDuck scan flagged Newtonsoft.Json 13.0.1 as a medium security risk (BDSA-2018-5195) because of the following issue:

Newtonsoft.Json is vulnerable to denial-of-service (DoS) due to a stack overflow that can occur whenever nested objects are being processed. A remote attacker could cause a vulnerable application to crash by causing it to process a maliciously crafted JSON object.